Create Anjana_v_m.md
Prime Numbers are those numbers which have 1 and that number itself as its factors.  In the function, we can take only natural numbers as input, not decimals or zero. That is what the first if loop does. If input is number other than natural numbers, We can say it's not a prime number and returns False. The for loop is used to iterate from range 2 to (n-1)th number. If any value of i can divide input, Then It is not a prime number and return False. If any value of i doesn't divide input, Then it's a prime number and return True.